fix(blog): unify spacing, media radius, and tag chips
- Drop the table-of-contents trailing item margin so the box padding alone
  sets the bottom gap; tighten and even out item spacing.
- Reset the prose-list gutter for the related-posts list and tag chips when
  they sit inside a `main section`, removing the stray left indent.
- Pad Shiki code blocks so code no longer touches the box edge, and give
  code, post featured images and in-content figures the shared media radius.
- Restyle tag chips as filled pills (no more faint border) and make tag
  clicks on the blog index and search results navigate to the tag archive
  instead of filling the search box.
- Extract a dependency-free `slugifyTag` so the client search bundle can
  build archive links without pulling in server-only modules.
