+++ src/components/PostListItem.astro
@@ -7,20 +7,14 @@ import { inlineMarkdown } from '../lib/inline-markdown';
 import { formatPolishDate } from '../lib/date';
 import { slugifyTag } from '../lib/tags';
 
-// Shared blog post card (one <li> in a .posts list). Used by the blog index and
-// the tag archives so the markup stays in one place.
-//
-// tagMode controls the tag chips:
-//   - 'link'   → real links to the tag archive (/tags/<slug>/); the default and
-//                the right choice everywhere except the blog index.
-//   - 'search' → the blog index's progressive-enhancement chips (data-tag drives
-//                the in-page search; href is the shareable /blog/?q= fallback).
+// Shared blog post card (one <li> in a .posts list). Used by the blog index, the
+// tag archives and the related-posts section so the markup stays in one place.
+// Tag chips always link to their archive page (/tags/<slug>/).
 interface Props {
   post: CollectionEntry<'posts'>;
-  tagMode?: 'link' | 'search';
 }
 
-const { post, tagMode = 'link' } = Astro.props;
+const { post } = Astro.props;
 const author = SITE_METADATA.author;
 
 const description = post.data.description || excerpt(post.body ?? '');
@@ -61,21 +55,13 @@ const dateFormatted = formatPolishDate(post.data.date);
     </section>
     <ul class="tags" aria-label="Tagi">
       {
-        post.data.tags.map(tag =>
-          tagMode === 'search' ? (
-            <li>
-              <a class="tag" href={`/blog/?q=${encodeURIComponent(tag)}`} data-tag={tag}>
-                #{tag}
-              </a>
-            </li>
-          ) : (
-            <li>
-              <a class="tag" href={`/tags/${slugifyTag(tag)}/`}>
-                #{tag}
-              </a>
-            </li>
-          ),
-        )
+        post.data.tags.map(tag => (
+          <li>
+            <a class="tag" href={`/tags/${slugifyTag(tag)}/`}>
+              #{tag}
+            </a>
+          </li>
+        ))
       }
     </ul>
   </article>
